Fixed major unlinking issues where applink would remove core symlinks such as /bin, /lib, etc..

+++ applib/fs.py
@@ -9,6 +9,7 @@
 import os
 import shutil
 import stat
+import sys
 from applib.general import AppType, AppFolders, MultipleVersionsException, NoVersionsException
 from applib.general import InvalidApplicationException, InvalidApplicationTypeException
 from applib.logging import default_logger as log
@@ -108,6 +109,79 @@ class InstalledApplication():
 				True
 				);
 		
+		safe_app_dir = os.path.join(
+                                        AppFolders.get(self.type),
+                                        self.name + "/" + self.version
+                                        )
+		
+		# Preemptively go through the list of directories, removing those
+		# that are symlinks to the application folder.  This is from the legacy
+		# link system and unfortunatly if you let the block below this run
+		# through a system with said symlinks, you'll end up annihilating the
+		# the application files (because it'll walk through the symlink into
+		# the application directory and start rm'ing stuff we don't want to)
+		# The solution here is to go through and remove directory symlinks before
+		# hand, with a reversed result list (in effect reversing the walk process
+		# in adiff.scan) so that we elimate the top level symlinks first, preventing
+		# it from annihilating symlinked directories inside the application folder.
+		# Very annoying stuff.
+		#
+		# XXX: I almost hosed the entire Elementary system with this.  Apparently it
+		#      that removing symlinked directories included some of the base ones
+		#      such as /lib and /bin (because the Python install contains those dirs
+		#      too :P).  The only_sub variable defines that only paths that resolve
+		#      to a *subdirectory* of those specified can be removed if it's a symlinked
+		#      directory.  This prevents removal of /bin, /lib, etc.. symlinks.
+		#
+		only_sub = [
+				"/System/Utilities/Applications",
+				"/System/Utilities/Libraries",
+				"/Applications",
+				"/Users"
+			]
+		results.reverse()
+		trip_safety = False
+		for i in results:
+			try:
+				pstat = os.lstat(i[2])[stat.ST_MODE]
+			except:
+				# Likely broken when we removed a directory symlink.
+				continue
+			
+			# Determine whether we should proceed with this entry.
+			if (not i[0] == "directory"):
+				continue
+			if (not stat.S_ISLNK(pstat)):
+				continue
+
+			# Determine whether it's safe to remove this symlinked dir.
+			if (not self.isApplicationOwned(i[2], safe_app_dir)):
+				log.showInfoW("Ignoring " + i[2] + " because it's not owned by the application (safety check).")
+				continue
+			
+			# Double-check before we go unlinking (in case of a logic oversight).
+			if (i[0] == "directory" and stat.S_ISLNK(pstat)):
+				trip_safety = True
+				try:
+					os.unlink(i[2])
+					log.showWarningW("Removed symlinked directory at: " + i[2])
+					log.showWarningW("The full path was: " + rpath)
+				except:
+					pass
+		results.reverse()		
+
+		if (trip_safety):
+			log.showErrorW("Legacy system safety switch was tripped.  This indicates you have")
+			log.showErrorO("symlinked directories on your system (from legacy linkage systems).")
+			log.showErrorO("The unlinking process has removed at least one of those symlinked")
+			log.showErrorO("directories.  In order to make sure application files don't get")
+			log.showErrorO("removed, you need to run the unlink process again to ensure the system")
+			log.showErrorO("is scanned without symlinked directories.  If the process shows this")
+			log.showErrorO("message twice, then STOP and REMOVE THE SYMLINKS MANUALLY.  You risk")
+			log.showErrorO("destroying application installations if you continue.")
+			sys.exit(1)
+		
+
 		# Now go through the results, removing directories (if they're
 		# empty) and un-symlinking files (but making sure that we only
 		# remove symlinks and not normal files).
@@ -117,7 +191,20 @@ class InstalledApplication():
 		total_files = 0
 		for i in results:
 			total_files += 1
-			if (i[0] == "directory"):
+			try:
+				pstat = os.lstat(i[2])[stat.ST_MODE]
+			except:
+				# File doesn't exist.  Likely got removed while we unlinked
+				# a `symlinked' directory (from old linkage system).
+				continue
+
+			# Check to make sure that the file we're going to remove is located
+			# within a safe directory.
+			if (not self.isApplicationOwned(i[2], safe_app_dir)):
+				log.showInfoW("Ignoring " + i[2] + " because it's not owned by the application.")
+				continue
+
+			if (i[0] == "directory" and not stat.S_ISLNK(pstat)):
 				try:
 					os.rmdir(i[2])
 					attempt_successes.append(i[2])
@@ -127,8 +214,9 @@ class InstalledApplication():
 					# as a failure since quite often directories will not be
 					# removed because they are still in use by other applications.
 					#attempt_failures.append(i[2])
-			elif (i[0] == "file" and stat.S_ISLNK(os.lstat(i[2])[stat.ST_MODE])):
+			elif ((i[0] == "file" or i[0] == "directory") and stat.S_ISLNK(pstat)):
 				try:
+					log.showInfoW("  * " + i[2])
 					os.unlink(i[2])
 					attempt_successes.append(i[2])
 				except:
@@ -136,7 +224,7 @@ class InstalledApplication():
 					attempt_failures.append(i[2])
 			elif (i[0] == "notexists"):
 				attempt_notexists.append(i[2])
-			else:
+			elif (i[0] != "notexists" and i[0] != "file" and i[0] != "directory"):
 				log.showWarningW("Unknown operation for " + i[1])
 
 		return attempt_successes, attempt_failures, total_files
@@ -154,6 +242,20 @@ class InstalledApplication():
 			AppFolders.get(self.type),
 			self.name + "/" + self.version
 			)
+	
+	"""Checks to see whether the specified path is located within safe_root
+	   when normalized."""
+	def isApplicationOwned(self, path, safe_root):
+		only_sub = [
+				safe_root
+                        ]
+		rpath = os.path.realpath(path)
+		issafe = False
+		for c in only_sub:
+			if (rpath.startswith(c) and rpath != c):
+				issafe = True
+		return issafe
+		
 
 class ApplicationDifferencer():
 	"""A class which determines the differences between what
